Move bounding box generation into computed on model
